pyproject.toml
[project]
name = "inventory-management-django-system"
version = "1.0.0"
description = "Inventory Management System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=0.9.9",
    "drf-spectacular[sidecar]>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "django-redis>=5.4.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "django-cors-headers>=4.3.1",
    "sentry-sdk[django]>=1.40.6",
    "gunicorn>=21.2.0",
    "uvicorn[standard]>=0.29.0",
    "flower>=2.0.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "app",
    "authentication",
    "brands",
    "categories",
    "inflows",
    "outflows",
    "product_models",
    "products",
    "suppliers",
]

# Dependências de Desenvolvimento (instaladas com uv sync --dev)
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.3",
    "time-machine>=2.13.0",
    "factory-boy>=3.3.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

# ========================================================
# CONFIGURAÇÃO DO RUFF (Linter & Formatter)
# ========================================================
[tool.ruff]
# Comprimento da linha (compatível com Black)
line-length = 88
# Versão alvo do Python
target-version = "py313"

[tool.ruff.lint]
# Habilita regras:
# E: pycodestyle errors
# F: Pyflakes
# I: isort (ordenação de imports)
# B: flake8-bugbear (bugs comuns)
# UP: pyupgrade (modernização de sintaxe)
select = ["E", "F", "I", "B", "UP"]
ignore = []

# Configurações específicas para Django
[tool.ruff.lint.isort]
known-first-party = ["app", "config"] # Suas pastas locais

# Ignorar regras específicas em arquivos de migração (são autogerados)
[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501"] # Ignora linha muito longa em migrações

# ========================================================
# CONFIGURAÇÃO DO MYPY (Static Type Checker)
# ========================================================
[tool.mypy]
python_version = ">=3.13"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Rigor na checagem
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true

# Ignorar pastas de migrações e testes na checagem estrita
exclude = [
    "manage.py",
    "**/migrations/",
    "**/tests/"
]

[tool.django-stubs]
django_settings_module = "app.settings"

# ========================================================
# CONFIGURAÇÃO DO PYTEST
# ========================================================
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["test_*.py", "*_tests.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]

# Coverage and Runtime options
addopts = "--reuse-db --nomigrations --strict-markers --tb=short --cov=app --cov=authentication --cov=brands --cov=categories --cov=inflows --cov=outflows --cov=products --cov=suppliers --cov-report=term-missing"

# Environment variables for testing (Requires pytest-env)
env = [
    "POSTGRES_DB=",  # Force SQLite for tests
    "POSTGRES_USER=",
    "POSTGRES_PASSWORD=",
    "POSTGRES_HOST=",
    "POSTGRES_PORT=",
]

# Custom markers
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "api: API endpoint tests",
    "signals: Signal handler tests",
    "slow: Tests that run slowly",
]

# Output
console_output_style = "progress"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
